refactor: rate_limitter.py > utils.py
